ssh: improved dialog (fixes #1093) (#1094)
