+++ app/src/main/java/io/treehouses/remote/Fragments/DialogFragments/HelpDialog.kt
@@ -6,17 +6,17 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.view.Window
-import androidx.fragment.app.DialogFragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import io.treehouses.remote.Constants
 import io.treehouses.remote.Views.RecyclerViewClickListener
 import io.treehouses.remote.adapter.HelpAdapter
+import io.treehouses.remote.bases.FullScreenDialogFragment
 import io.treehouses.remote.databinding.DialogHelpBinding
 import io.treehouses.remote.pojo.HelpCommand
 import org.json.JSONObject
 
 
-class HelpDialog : DialogFragment(), android.widget.SearchView.OnQueryTextListener {
+class HelpDialog : FullScreenDialogFragment(), android.widget.SearchView.OnQueryTextListener {
     private lateinit var bind: DialogHelpBinding
     private var jsonString = ""
     private val items = mutableListOf<HelpCommand>()
@@ -74,18 +74,6 @@ class HelpDialog : DialogFragment(), android.widget.SearchView.OnQueryTextListen
         bind.showHelp.visibility = View.VISIBLE
         bind.titleDescription.text = item.title
         bind.description.text = item.preview
-    }
-
-
-    override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
-        val dialog: Dialog = super.onCreateDialog(savedInstanceState)
-        dialog.requestWindowFeature(Window.FEATURE_NO_TITLE)
-        return dialog
-    }
-
-    override fun onStart() {
-        super.onStart()
-        dialog?.window?.setLayout(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT)
     }
 
     private fun createJson(jsonStr: String) {
