+++ app/src/main/java/io/treehouses/remote/Fragments/DialogFragments/SSHDialog.kt
@@ -1,29 +1,35 @@
 package io.treehouses.remote.Fragments.DialogFragments
 
-import android.app.AlertDialog
 import android.content.Intent
+import android.os.Bundle
 import android.text.Editable
 import android.text.TextWatcher
-import androidx.appcompat.app.AppCompatActivity
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.FragmentActivity
 import io.treehouses.remote.Fragments.SSHConsole
-import io.treehouses.remote.R
 import io.treehouses.remote.SSH.beans.HostBean
+import io.treehouses.remote.bases.FullScreenDialogFragment
 import io.treehouses.remote.databinding.DialogSshBinding
 import java.util.regex.Pattern
 
-class SSHDialog(activity: AppCompatActivity) {
+
+class SSHDialog : FullScreenDialogFragment() {
     private val sshPattern = Pattern.compile("^(.+)@(([0-9a-z.-]+)|(\\[[a-f:0-9]+\\]))(:(\\d+))?$", Pattern.CASE_INSENSITIVE)
-    private var bind: DialogSshBinding = DialogSshBinding.inflate(activity.layoutInflater)
-    private var dialog: AlertDialog? = null
-    init {
+    private lateinit var bind: DialogSshBinding
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        bind = DialogSshBinding.inflate(inflater, container, false)
+        return bind.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
         setEnabled(false)
         addTextValidation()
         bind.connectSsh.setOnClickListener {
-            launchSSH(activity, bind.sshTextInput.text.toString().split("@")[0], bind.sshTextInput.text.toString().split("@")[1])
-            dialog?.dismiss()
+            launchSSH(requireActivity(), bind.sshTextInput.text.toString().split("@")[0], bind.sshTextInput.text.toString().split("@")[1])
         }
-        dialog = AlertDialog.Builder(activity).setTitle("Start SSH Connection").setView(bind.root).setIcon(R.drawable.dialog_icon).create()
-        dialog?.show()
     }
 
     private fun addTextValidation() {
@@ -49,7 +55,7 @@ class SSHDialog(activity: AppCompatActivity) {
         bind.connectSsh.isClickable = bool
     }
 
-    private fun launchSSH(activity: AppCompatActivity, username: String, hostname: String) {
+    private fun launchSSH(activity: FragmentActivity, username: String, hostname: String) {
         val host = HostBean()
         host.username = username
         host.hostname = hostname
@@ -57,5 +63,6 @@ class SSHDialog(activity: AppCompatActivity) {
         contents.flags = Intent.FLAG_ACTIVITY_CLEAR_TOP
         contents.setClass(activity, SSHConsole::class.java)
         activity.startActivity(contents)
+        dialog?.dismiss()
     }
 }
