+++ app/src/main/res/drawable/rectangle_border.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="rectangle">
-    <stroke android:width="0.35dp" android:color="@color/md_grey_800"/>
+    <stroke android:width="0.35dp" android:color="@color/rectangle_border_color"/>
     <corners android:radius="4dp"/>
 </shape>
