+++ app/src/main/res/layout/dialog_ssh.xml
@@ -3,15 +3,26 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:orientation="vertical"
+    android:background="@color/card_background"
     xmlns:app="http://schemas.android.com/apk/res-auto">
-
+    <TextView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="Start SSH Connection"
+        android:textColor="@color/daynight_textColor"
+        android:paddingStart="@dimen/padding_normal"
+        android:paddingEnd="@dimen/padding_normal"
+        android:paddingTop="@dimen/padding_normal"
+        android:gravity="center"
+        android:textAppearance="@style/TextAppearance.AppCompat.Title"/>
     <com.google.android.material.textfield.TextInputEditText
         android:id="@+id/ssh_text_input"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
+        android:layout_height="wrap_content"
         android:background="@drawable/rectangle_border"
         android:padding="@dimen/padding_normal"
         android:layout_margin="@dimen/margin_large"
+        android:textColorHint="@color/md_grey_500"
         android:hint="pi@192.168.1.29"
         />
