+++ app/src/main/res/values-night/color.xml
@@ -25,6 +25,8 @@
     <color name="terminal">#FFFFFF</color>
     <color name="back_arrow">#CCCCCC</color>
 
+    <color name="rectangle_border_color">@color/md_grey_500</color>
+
     <color name="black_overlay">#66000000</color>
     <color name="bg_white">#ffffff</color>
     <color name="red">#66ff0000</color>
